+++ CHANGELOG.md
@@ -6,6 +6,7 @@ Release Notes
 
 * **NEW:** Support for outputting teletext packets directly to standard output
 * **NEW:** Support for outputting teletext packets to a web socket
+* **NEW:** Support for outputting teletext packets as a T42 file
 * **NEW:** Support for looped output
 * **IMPROVED:** The code has been significantly refactored for improved readability
 * **FIXED:** Out of bounds crash that occurred while decoding enhancement packets